+++ client/src/containers/App.jsx
@@ -1,9 +1,9 @@
 import { useState } from 'react'
-import reactLogo from './assets/react.svg'
+import reactLogo from '.././assets/react.svg'
 import viteLogo from '/vite.svg'
-import './App.css'
+import '.././assets/App.css'
 import { Route, Routes } from 'react-router-dom';
-import Login from './components/Login';
+import Login from '../components/Login';
 function App() {
 
   return (