+++ CHANGELOG.md
@@ -7,4 +7,4 @@ All notable changes to this project will be documented in this file. See [standa
 
 ### Features
 
-* add Signup and Login feature ([bc0d970](https://github.com/Nishant040305/vite/commit/bc0d9709d9028bf386e6d2e0148b73cf9754bc94))
+* add Signup and Login feature ([bc0d970](https://github.com/Nishant040305/HorizanArcInvestment/commit/bc0d9709d9028bf386e6d2e0148b73cf9754bc94))